requirements.txt
nextcord
typing
asyncio
requests
pillow
urllib3
psutil
python-dotenv
validators
wikipedia
datetime
easy-pil
aiomojang
httpx
nextcord[voice]
bs4
humanfriendly
timeago
aioconsole
mctools